requirements.txt
jinja2
web.py
pyserial
